+++ core/cmd/evm_node_commands_test.go
@@ -27,16 +27,22 @@ func TestClient_IndexEVMNodes(t *testing.T) {
 	t.Parallel()
 
 	chainID := newRandChainID()
-	node := evmcfg.Node{
-		Name:     ptr("Test node"),
+	node1 := evmcfg.Node{
+		Name:     ptr("Test node 1"),
 		WSURL:    models.MustParseURL("ws://localhost:8546"),
 		HTTPURL:  models.MustParseURL("http://localhost:8546"),
+		SendOnly: ptr(false),
+	}
+	node2 := evmcfg.Node{
+		Name:     ptr("Test node 2"),
+		WSURL:    models.MustParseURL("ws://localhost:8547"),
+		HTTPURL:  models.MustParseURL("http://localhost:8547"),
 		SendOnly: ptr(false),
 	}
 	chain := evmcfg.EVMConfig{
 		ChainID: chainID,
 		Chain:   evmcfg.Defaults(chainID),
-		Nodes:   evmcfg.EVMNodes{&node},
+		Nodes:   evmcfg.EVMNodes{&node1, &node2},
 	}
 	app := startNewApplicationV2(t, func(c *chainlink.Config, s *chainlink.Secrets) {
 		c.EVM = evmcfg.EVMConfigs{&chain}
@@ -46,11 +52,16 @@ func TestClient_IndexEVMNodes(t *testing.T) {
 	require.Nil(t, cmd.NewEVMNodeClient(client).IndexNodes(cltest.EmptyCLIContext()))
 	require.NotEmpty(t, r.Renders)
 	nodes := *r.Renders[0].(*cmd.EVMNodePresenters)
-	require.Len(t, nodes, 1)
-	n := nodes[0]
-	assert.Equal(t, "0", n.ID)
-	assert.Equal(t, *node.Name, n.Name)
-	assert.Equal(t, node.WSURL.String(), n.WSURL.String)
-	assert.Equal(t, node.HTTPURL.String(), n.HTTPURL.String)
+	require.Len(t, nodes, 2)
+	n1 := nodes[0]
+	n2 := nodes[1]
+	assert.Equal(t, "Test node 1", n1.ID)
+	assert.Equal(t, *node1.Name, n1.Name)
+	assert.Equal(t, node1.WSURL.String(), n1.WSURL.String)
+	assert.Equal(t, node1.HTTPURL.String(), n1.HTTPURL.String)
+	assert.Equal(t, "Test node 2", n2.ID)
+	assert.Equal(t, *node2.Name, n2.Name)
+	assert.Equal(t, node2.WSURL.String(), n2.WSURL.String)
+	assert.Equal(t, node2.HTTPURL.String(), n2.HTTPURL.String)
 	assertTableRenders(t, r)
 }