+++ integration-tests/contracts/contract_deployer.go
@@ -6,8 +6,6 @@ import (
 	"math/big"
 	"time"
 
-	int_ethereum "github.com/smartcontractkit/chainlink/integration-tests/contracts/ethereum"
-
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
@@ -19,7 +17,17 @@ import (
 	"github.com/smartcontractkit/chainlink-testing-framework/blockchain"
 	"github.com/smartcontractkit/chainlink-testing-framework/contracts/ethereum"
 
-	ethereum2 "github.com/smartcontractkit/chainlink/integration-tests/contracts/ethereum"
+	eth_contracts "github.com/smartcontractkit/chainlink/integration-tests/contracts/ethereum"
+
+	"github.com/smartcontractkit/chainlink/core/gethwrappers/generated/keeper_registrar_wrapper1_2"
+	"github.com/smartcontractkit/chainlink/core/gethwrappers/generated/keeper_registrar_wrapper2_0"
+	"github.com/smartcontractkit/chainlink/core/gethwrappers/generated/keeper_registry_logic1_3"
+	"github.com/smartcontractkit/chainlink/core/gethwrappers/generated/keeper_registry_logic2_0"
+	"github.com/smartcontractkit/chainlink/core/gethwrappers/generated/keeper_registry_wrapper1_1"
+	"github.com/smartcontractkit/chainlink/core/gethwrappers/generated/keeper_registry_wrapper1_2"
+	"github.com/smartcontractkit/chainlink/core/gethwrappers/generated/keeper_registry_wrapper1_3"
+	"github.com/smartcontractkit/chainlink/core/gethwrappers/generated/keeper_registry_wrapper2_0"
+	"github.com/smartcontractkit/chainlink/core/gethwrappers/generated/upkeep_transcoder"
 )
 
 // ContractDeployer is an interface for abstracting the contract deployment methods across network implementations
@@ -39,7 +47,7 @@ type ContractDeployer interface {
 	DeployVRFContract() (VRF, error)
 	DeployMockETHLINKFeed(answer *big.Int) (MockETHLINKFeed, error)
 	DeployMockGasFeed(answer *big.Int) (MockGasFeed, error)
-	DeployKeeperRegistrar(registryVersion ethereum.KeeperRegistryVersion, linkAddr string, registrarSettings KeeperRegistrarSettings) (KeeperRegistrar, error)
+	DeployKeeperRegistrar(registryVersion eth_contracts.KeeperRegistryVersion, linkAddr string, registrarSettings KeeperRegistrarSettings) (KeeperRegistrar, error)
 	DeployUpkeepTranscoder() (UpkeepTranscoder, error)
 	DeployKeeperRegistry(opts *KeeperRegistryOpts) (KeeperRegistry, error)
 	DeployKeeperConsumer(updateInterval *big.Int) (KeeperConsumer, error)
@@ -72,7 +80,7 @@ type ContractDeployer interface {
 	DeployBlockhashStore() (BlockHashStore, error)
 	DeployOperatorFactory(linkAddr string) (OperatorFactory, error)
 	DeployUpkeepResetter() (UpkeepResetter, error)
-	DeployStaking(params ethereum2.StakingPoolConstructorParams) (Staking, error)
+	DeployStaking(params eth_contracts.StakingPoolConstructorParams) (Staking, error)
 	DeployBatchBlockhashStore(blockhashStoreAddr string) (BatchBlockhashStore, error)
 	DeployAtlasFunctions() (AtlasFunctions, error)
 	LoadVerifierProxy(address common.Address) (VerifierProxy, error)
@@ -255,19 +263,19 @@ func (e *EthereumContractDeployer) DeployFluxAggregatorContract(
 	}, nil
 }
 
-func (e *EthereumContractDeployer) DeployStaking(params ethereum2.StakingPoolConstructorParams) (Staking, error) {
+func (e *EthereumContractDeployer) DeployStaking(params eth_contracts.StakingPoolConstructorParams) (Staking, error) {
 	stakingAddress, _, instance, err := e.client.DeployContract("Staking", func(
 		auth *bind.TransactOpts,
 		backend bind.ContractBackend,
 	) (common.Address, *types.Transaction, interface{}, error) {
-		return ethereum2.DeployStaking(auth, backend, params)
+		return eth_contracts.DeployStaking(auth, backend, params)
 	})
 	if err != nil {
 		return nil, err
 	}
 	return &EthereumStaking{
 		client:  e.client,
-		staking: instance.(*ethereum2.Staking),
+		staking: instance.(*eth_contracts.Staking),
 		address: stakingAddress,
 	}, nil
 }
@@ -277,14 +285,14 @@ func (e *EthereumContractDeployer) DeployAtlasFunctions() (AtlasFunctions, error
 		auth *bind.TransactOpts,
 		backend bind.ContractBackend,
 	) (common.Address, *types.Transaction, interface{}, error) {
-		return ethereum2.DeployAtlasFunctions(auth, backend)
+		return eth_contracts.DeployAtlasFunctions(auth, backend)
 	})
 	if err != nil {
 		return nil, err
 	}
 	return &EthereumAtlasFunctions{
 		client:         e.client,
-		atlasFunctions: instance.(*ethereum2.AtlasFunctions),
+		atlasFunctions: instance.(*eth_contracts.AtlasFunctions),
 		address:        address,
 	}, nil
 }
@@ -478,7 +486,7 @@ func (e *EthereumContractDeployer) DeployUpkeepTranscoder() (UpkeepTranscoder, e
 		opts *bind.TransactOpts,
 		backend bind.ContractBackend,
 	) (common.Address, *types.Transaction, interface{}, error) {
-		return ethereum.DeployUpkeepTranscoder(opts, backend)
+		return upkeep_transcoder.DeployUpkeepTranscoder(opts, backend)
 	})
 
 	if err != nil {
@@ -487,21 +495,21 @@ func (e *EthereumContractDeployer) DeployUpkeepTranscoder() (UpkeepTranscoder, e
 
 	return &EthereumUpkeepTranscoder{
 		client:     e.client,
-		transcoder: instance.(*ethereum.UpkeepTranscoder),
+		transcoder: instance.(*upkeep_transcoder.UpkeepTranscoder),
 		address:    address,
 	}, err
 }
 
-func (e *EthereumContractDeployer) DeployKeeperRegistrar(registryVersion ethereum.KeeperRegistryVersion, linkAddr string,
+func (e *EthereumContractDeployer) DeployKeeperRegistrar(registryVersion eth_contracts.KeeperRegistryVersion, linkAddr string,
 	registrarSettings KeeperRegistrarSettings) (KeeperRegistrar, error) {
 
-	if registryVersion == ethereum.RegistryVersion_2_0 {
+	if registryVersion == eth_contracts.RegistryVersion_2_0 {
 		// deploy registrar 2.0
 		address, _, instance, err := e.client.DeployContract("KeeperRegistrar", func(
 			opts *bind.TransactOpts,
 			backend bind.ContractBackend,
 		) (common.Address, *types.Transaction, interface{}, error) {
-			return ethereum.DeployKeeperRegistrar20(opts, backend, common.HexToAddress(linkAddr), registrarSettings.AutoApproveConfigType,
+			return keeper_registrar_wrapper2_0.DeployKeeperRegistrar(opts, backend, common.HexToAddress(linkAddr), registrarSettings.AutoApproveConfigType,
 				registrarSettings.AutoApproveMaxAllowed, common.HexToAddress(registrarSettings.RegistryAddr), registrarSettings.MinLinkJuels)
 		})
 
@@ -511,7 +519,7 @@ func (e *EthereumContractDeployer) DeployKeeperRegistrar(registryVersion ethereu
 
 		return &EthereumKeeperRegistrar{
 			client:      e.client,
-			registrar20: instance.(*ethereum.KeeperRegistrar20),
+			registrar20: instance.(*keeper_registrar_wrapper2_0.KeeperRegistrar),
 			address:     address,
 		}, err
 	}
@@ -520,7 +528,7 @@ func (e *EthereumContractDeployer) DeployKeeperRegistrar(registryVersion ethereu
 		opts *bind.TransactOpts,
 		backend bind.ContractBackend,
 	) (common.Address, *types.Transaction, interface{}, error) {
-		return ethereum.DeployKeeperRegistrar(opts, backend, common.HexToAddress(linkAddr), registrarSettings.AutoApproveConfigType,
+		return keeper_registrar_wrapper1_2.DeployKeeperRegistrar(opts, backend, common.HexToAddress(linkAddr), registrarSettings.AutoApproveConfigType,
 			registrarSettings.AutoApproveMaxAllowed, common.HexToAddress(registrarSettings.RegistryAddr), registrarSettings.MinLinkJuels)
 	})
 
@@ -530,7 +538,7 @@ func (e *EthereumContractDeployer) DeployKeeperRegistrar(registryVersion ethereu
 
 	return &EthereumKeeperRegistrar{
 		client:    e.client,
-		registrar: instance.(*ethereum.KeeperRegistrar),
+		registrar: instance.(*keeper_registrar_wrapper1_2.KeeperRegistrar),
 		address:   address,
 	}, err
 }
@@ -538,25 +546,25 @@ func (e *EthereumContractDeployer) DeployKeeperRegistrar(registryVersion ethereu
 func (e *EthereumContractDeployer) DeployKeeperRegistry(
 	opts *KeeperRegistryOpts,
 ) (KeeperRegistry, error) {
-	var paymentModel uint8
+	var mode uint8
 	switch e.client.GetChainID() {
 	//Arbitrum payment model
 	case big.NewInt(421613):
-		paymentModel = uint8(1)
+		mode = uint8(1)
 	//Optimism payment model
 	case big.NewInt(420):
-		paymentModel = uint8(2)
+		mode = uint8(2)
 	default:
-		paymentModel = uint8(0)
+		mode = uint8(0)
 	}
 	registryGasOverhead := big.NewInt(80000)
 	switch opts.RegistryVersion {
-	case ethereum.RegistryVersion_1_0, ethereum.RegistryVersion_1_1:
+	case eth_contracts.RegistryVersion_1_0, eth_contracts.RegistryVersion_1_1:
 		address, _, instance, err := e.client.DeployContract("KeeperRegistry1_1", func(
 			auth *bind.TransactOpts,
 			backend bind.ContractBackend,
 		) (common.Address, *types.Transaction, interface{}, error) {
-			return ethereum.DeployKeeperRegistry11(
+			return keeper_registry_wrapper1_1.DeployKeeperRegistry(
 				auth,
 				backend,
 				common.HexToAddress(opts.LinkAddr),
@@ -577,24 +585,24 @@ func (e *EthereumContractDeployer) DeployKeeperRegistry(
 		}
 		return &EthereumKeeperRegistry{
 			client:      e.client,
-			version:     ethereum.RegistryVersion_1_1,
-			registry1_1: instance.(*ethereum.KeeperRegistry11),
+			version:     eth_contracts.RegistryVersion_1_1,
+			registry1_1: instance.(*keeper_registry_wrapper1_1.KeeperRegistry),
 			registry1_2: nil,
 			registry1_3: nil,
 			address:     address,
 		}, err
-	case ethereum.RegistryVersion_1_2:
+	case eth_contracts.RegistryVersion_1_2:
 		address, _, instance, err := e.client.DeployContract("KeeperRegistry1_2", func(
 			auth *bind.TransactOpts,
 			backend bind.ContractBackend,
 		) (common.Address, *types.Transaction, interface{}, error) {
-			return ethereum.DeployKeeperRegistry12(
+			return keeper_registry_wrapper1_2.DeployKeeperRegistry(
 				auth,
 				backend,
 				common.HexToAddress(opts.LinkAddr),
 				common.HexToAddress(opts.ETHFeedAddr),
 				common.HexToAddress(opts.GasFeedAddr),
-				ethereum.Config1_2{
+				keeper_registry_wrapper1_2.Config{
 					PaymentPremiumPPB:    opts.Settings.PaymentPremiumPPB,
 					FlatFeeMicroLink:     opts.Settings.FlatFeeMicroLINK,
 					BlockCountPerTurn:    opts.Settings.BlockCountPerTurn,
@@ -615,21 +623,21 @@ func (e *EthereumContractDeployer) DeployKeeperRegistry(
 		}
 		return &EthereumKeeperRegistry{
 			client:      e.client,
-			version:     ethereum.RegistryVersion_1_2,
+			version:     eth_contracts.RegistryVersion_1_2,
 			registry1_1: nil,
-			registry1_2: instance.(*ethereum.KeeperRegistry12),
+			registry1_2: instance.(*keeper_registry_wrapper1_2.KeeperRegistry),
 			registry1_3: nil,
 			address:     address,
 		}, err
-	case ethereum.RegistryVersion_1_3:
+	case eth_contracts.RegistryVersion_1_3:
 		logicAddress, _, _, err := e.client.DeployContract("KeeperRegistryLogic1_3", func(
 			auth *bind.TransactOpts,
 			backend bind.ContractBackend,
 		) (common.Address, *types.Transaction, interface{}, error) {
-			return ethereum.DeployKeeperRegistryLogic13(
+			return keeper_registry_logic1_3.DeployKeeperRegistryLogic(
 				auth,
 				backend,
-				paymentModel,        // Default payment model
+				mode,                // Default payment model
 				registryGasOverhead, // Registry gas overhead
 				common.HexToAddress(opts.LinkAddr),
 				common.HexToAddress(opts.ETHFeedAddr),
@@ -648,11 +656,11 @@ func (e *EthereumContractDeployer) DeployKeeperRegistry(
 			auth *bind.TransactOpts,
 			backend bind.ContractBackend,
 		) (common.Address, *types.Transaction, interface{}, error) {
-			return ethereum.DeployKeeperRegistry13(
+			return keeper_registry_wrapper1_3.DeployKeeperRegistry(
 				auth,
 				backend,
 				*logicAddress,
-				ethereum.Config1_3{
+				keeper_registry_wrapper1_3.Config{
 					PaymentPremiumPPB:    opts.Settings.PaymentPremiumPPB,
 					FlatFeeMicroLink:     opts.Settings.FlatFeeMicroLINK,
 					BlockCountPerTurn:    opts.Settings.BlockCountPerTurn,
@@ -673,21 +681,21 @@ func (e *EthereumContractDeployer) DeployKeeperRegistry(
 		}
 		return &EthereumKeeperRegistry{
 			client:      e.client,
-			version:     ethereum.RegistryVersion_1_3,
+			version:     eth_contracts.RegistryVersion_1_3,
 			registry1_1: nil,
 			registry1_2: nil,
-			registry1_3: instance.(*ethereum.KeeperRegistry13),
+			registry1_3: instance.(*keeper_registry_wrapper1_3.KeeperRegistry),
 			address:     address,
 		}, err
-	case ethereum.RegistryVersion_2_0:
+	case eth_contracts.RegistryVersion_2_0:
 		logicAddress, _, _, err := e.client.DeployContract("KeeperRegistryLogic2_0", func(
 			auth *bind.TransactOpts,
 			backend bind.ContractBackend,
 		) (common.Address, *types.Transaction, interface{}, error) {
-			return ethereum.DeployKeeperRegistryLogic20(
+			return keeper_registry_logic2_0.DeployKeeperRegistryLogic(
 				auth,
 				backend,
-				paymentModel, // Default payment model
+				mode, // Default payment model
 				common.HexToAddress(opts.LinkAddr),
 				common.HexToAddress(opts.ETHFeedAddr),
 				common.HexToAddress(opts.GasFeedAddr),
@@ -706,7 +714,7 @@ func (e *EthereumContractDeployer) DeployKeeperRegistry(
 			backend bind.ContractBackend,
 		) (common.Address, *types.Transaction, interface{}, error) {
 
-			return ethereum.DeployKeeperRegistry20(
+			return keeper_registry_wrapper2_0.DeployKeeperRegistry(
 				auth,
 				backend,
 				*logicAddress,
@@ -717,8 +725,8 @@ func (e *EthereumContractDeployer) DeployKeeperRegistry(
 		}
 		return &EthereumKeeperRegistry{
 			client:      e.client,
-			version:     ethereum.RegistryVersion_2_0,
-			registry2_0: instance.(*ethereum.KeeperRegistry20),
+			version:     eth_contracts.RegistryVersion_2_0,
+			registry2_0: instance.(*keeper_registry_wrapper2_0.KeeperRegistry),
 			address:     address,
 		}, err
 
@@ -732,14 +740,14 @@ func (e *EthereumContractDeployer) DeployKeeperConsumer(updateInterval *big.Int)
 		auth *bind.TransactOpts,
 		backend bind.ContractBackend,
 	) (common.Address, *types.Transaction, interface{}, error) {
-		return ethereum.DeployKeeperConsumer(auth, backend, updateInterval)
+		return eth_contracts.DeployKeeperConsumer(auth, backend, updateInterval)
 	})
 	if err != nil {
 		return nil, err
 	}
 	return &EthereumKeeperConsumer{
 		client:   e.client,
-		consumer: instance.(*ethereum.KeeperConsumer),
+		consumer: instance.(*eth_contracts.KeeperConsumer),
 		address:  address,
 	}, err
 }
@@ -749,14 +757,14 @@ func (e *EthereumContractDeployer) DeployUpkeepCounter(testRange *big.Int, inter
 		auth *bind.TransactOpts,
 		backend bind.ContractBackend,
 	) (common.Address, *types.Transaction, interface{}, error) {
-		return ethereum.DeployUpkeepCounter(auth, backend, testRange, interval)
+		return eth_contracts.DeployUpkeepCounter(auth, backend, testRange, interval)
 	})
 	if err != nil {
 		return nil, err
 	}
 	return &EthereumUpkeepCounter{
 		client:   e.client,
-		consumer: instance.(*ethereum.UpkeepCounter),
+		consumer: instance.(*eth_contracts.UpkeepCounter),
 		address:  address,
 	}, err
 }
@@ -766,14 +774,14 @@ func (e *EthereumContractDeployer) DeployUpkeepPerformCounterRestrictive(testRan
 		auth *bind.TransactOpts,
 		backend bind.ContractBackend,
 	) (common.Address, *types.Transaction, interface{}, error) {
-		return ethereum.DeployUpkeepPerformCounterRestrictive(auth, backend, testRange, averageEligibilityCadence)
+		return eth_contracts.DeployUpkeepPerformCounterRestrictive(auth, backend, testRange, averageEligibilityCadence)
 	})
 	if err != nil {
 		return nil, err
 	}
 	return &EthereumUpkeepPerformCounterRestrictive{
 		client:   e.client,
-		consumer: instance.(*ethereum.UpkeepPerformCounterRestrictive),
+		consumer: instance.(*eth_contracts.UpkeepPerformCounterRestrictive),
 		address:  address,
 	}, err
 }
@@ -788,7 +796,7 @@ func (e *EthereumContractDeployer) DeployKeeperConsumerPerformance(
 		auth *bind.TransactOpts,
 		backend bind.ContractBackend,
 	) (common.Address, *types.Transaction, interface{}, error) {
-		return ethereum.DeployKeeperConsumerPerformance(
+		return eth_contracts.DeployKeeperConsumerPerformance(
 			auth,
 			backend,
 			testBlockRange,
@@ -802,7 +810,7 @@ func (e *EthereumContractDeployer) DeployKeeperConsumerPerformance(
 	}
 	return &EthereumKeeperConsumerPerformance{
 		client:   e.client,
-		consumer: instance.(*ethereum.KeeperConsumerPerformance),
+		consumer: instance.(*eth_contracts.KeeperConsumerPerformance),
 		address:  address,
 	}, err
 }
@@ -818,7 +826,7 @@ func (e *EthereumContractDeployer) DeployKeeperConsumerBenchmark(
 		auth *bind.TransactOpts,
 		backend bind.ContractBackend,
 	) (common.Address, *types.Transaction, interface{}, error) {
-		return ethereum.DeployKeeperConsumerBenchmark(
+		return eth_contracts.DeployKeeperConsumerBenchmark(
 			auth,
 			backend,
 			testBlockRange,
@@ -833,7 +841,7 @@ func (e *EthereumContractDeployer) DeployKeeperConsumerBenchmark(
 	}
 	return &EthereumKeeperConsumerBenchmark{
 		client:   e.client,
-		consumer: instance.(*ethereum.KeeperConsumerBenchmark),
+		consumer: instance.(*eth_contracts.KeeperConsumerBenchmark),
 		address:  address,
 	}, err
 }
@@ -843,7 +851,7 @@ func (e *EthereumContractDeployer) DeployKeeperPerformDataChecker(expectedData [
 		auth *bind.TransactOpts,
 		backend bind.ContractBackend,
 	) (common.Address, *types.Transaction, interface{}, error) {
-		return ethereum.DeployPerformDataChecker(
+		return eth_contracts.DeployPerformDataChecker(
 			auth,
 			backend,
 			expectedData,
@@ -854,7 +862,7 @@ func (e *EthereumContractDeployer) DeployKeeperPerformDataChecker(expectedData [
 	}
 	return &EthereumKeeperPerformDataCheckerConsumer{
 		client:             e.client,
-		performDataChecker: instance.(*ethereum.PerformDataChecker),
+		performDataChecker: instance.(*eth_contracts.PerformDataChecker),
 		address:            address,
 	}, err
 }
@@ -883,7 +891,7 @@ func (e *EthereumContractDeployer) DeployUpkeepResetter() (UpkeepResetter, error
 		auth *bind.TransactOpts,
 		backend bind.ContractBackend,
 	) (common.Address, *types.Transaction, interface{}, error) {
-		return int_ethereum.DeployUpkeepResetter(auth, backend)
+		return eth_contracts.DeployUpkeepResetter(auth, backend)
 	})
 	if err != nil {
 		return nil, err
@@ -891,6 +899,6 @@ func (e *EthereumContractDeployer) DeployUpkeepResetter() (UpkeepResetter, error
 	return &EthereumUpkeepResetter{
 		address:  addr,
 		client:   e.client,
-		consumer: instance.(*int_ethereum.UpkeepResetter),
+		consumer: instance.(*eth_contracts.UpkeepResetter),
 	}, err
 }