+++ docs/CHANGELOG.md
@@ -8,13 +8,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- unreleased -->
 ## [dev]
 
+### Changed
+
+- TOML configuration and secrets are now scoped to `chainlink node start` command rather than being global flags
+
 ### Removed
+
 - Configuration with legacy environment variables is no longer supported. TOML is required.
 
 ## 1.13.0 - UNRELEASED
 
 ### Added
 
+- Support for sending Bootstrap job specs to the feeds manager
 - Support for sending OCR2 job specs to the feeds manager
 - Log poller filters now saved in db, restored on node startup to guard against missing logs during periods where services are temporarily unable to start
 - Add support for new job type `mercury` (low-latency oracle)